Split SQLdb into TrajDB and AMSDB.
